rename: lounge -> main 디렉터리 변경
