+++ client/src/components/CreatedEventList/index.tsx
@@ -1,5 +1,5 @@
 import useRequestDeleteEvents from '@hooks/queries/event/useRequestDeleteEvents';
-import {useCreatedEventsPageContext} from '@pages/lounge/events/CreatedEvent.context';
+import {useCreatedEventsPageContext} from '@pages/main/events/CreatedEvent.context';
 import {CreatedEvent} from 'types/serviceType';
 import {CreatedEventItem} from '@components/Design/components/CreatedEventItem/CreatedEventItem';
